demp,diffractive_vm: MCParticles.momentum is now double, not float

+++ benchmarks/Exclusive-Diffraction-Tagging/demp/analysis/demp_analysis.cxx
@@ -66,9 +66,9 @@ void demp_analysis(const std::string& config_name)
 
   // Get Particle Information
   TTreeReaderArray<int>    partGenStat(tree_reader, "MCParticles.generatorStatus");
-  TTreeReaderArray<float>  partMomX(tree_reader, "MCParticles.momentum.x");
-  TTreeReaderArray<float>  partMomY(tree_reader, "MCParticles.momentum.y");
-  TTreeReaderArray<float>  partMomZ(tree_reader, "MCParticles.momentum.z");
+  TTreeReaderArray<double> partMomX(tree_reader, "MCParticles.momentum.x");
+  TTreeReaderArray<double> partMomY(tree_reader, "MCParticles.momentum.y");
+  TTreeReaderArray<double> partMomZ(tree_reader, "MCParticles.momentum.z");
   TTreeReaderArray<int>    partPdg(tree_reader, "MCParticles.PDG");
   TTreeReaderArray<double> partMass(tree_reader,"MCParticles.mass");
